Extract service calls into another file
