+++ src/rpc_callbacks.h
@@ -1,0 +1,11 @@
+#ifndef RPC_CALLBACKS_H
+#define RPC_CALLBACKS_H
+
+#include <rpc_callbacks.h>
+#include <simplerpc/service_call.h>
+
+extern service_call_method service_call_callbacks[];
+
+extern const unsigned int service_call_callbacks_len;
+
+#endif
